# api/openapi/access-service.yaml
openapi: 3.0.3
info:
  title: OpenIDX Access Service
  description: |
    Zero Trust access proxy with route management, session control, OpenZiti network overlay, posture checks, BrowZer, Guacamole integration, and Privileged Access Management (vault-backed entries, brokered launch with credential injection, checkout controls, session moderation, SSH CA, cloud JIT, Windows app delivery).

    Every route cmd/access-service mounts appears here, and nothing that is not
    mounted does: test/openapi/spec_census_test.go compares this file against the
    live route table in both directions. Operations carry an `x-openidx-handler`
    extension naming the Go handler that serves them; where a payload shape is not
    yet pinned the body is shown as a free-form JSON object rather than invented.
  version: 1.34.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8007
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/access/agent/enroll:
    post:
      tags: [Agent Enrollment]
      summary: Handle enroll
      operationId: handleEnroll
      x-openidx-handler: internal/access.(*AgentAPIHandler).HandleEnroll
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/v1/access/enroll:
    post:
      tags: [Agent Enrollment]
      summary: Enroll
      operationId: enroll
      x-openidx-handler: internal/access.(*Service).handleEnroll
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: The request could not be completed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: A short machine-readable reason
        error_description:
          type: string
          description: Operator-facing detail, when the handler has one
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

    # --- Auth ---
...